norl_week_1_complete
This is the final code for week 1 of programming (and week 5 of the capstone project). It thas basic page structure, with sober CSS, MongoDB interactions and user can modify data. User inputs and date wise filtering of todos is something I want to focus on in the next week.

create a new meteor project and replace everything other than .meteor folder with these files or download the zip file, unzip it and you're ready to go.